Implemented Operator Frenzy game, using a queue data structure for ordering arithmetic operators.
Operator overloading is used for easily arithmetic performance on each item pulled from the queue.

Utility class created for Operator Frenzy game to assist with these operations.

+++ Casino Game/Casino Game/Player.h
@@ -18,7 +18,7 @@ class Player
 private:
 	std::string playerName;
 	int playerCash;
-	int playerScore;
+	long int playerScore;
 	int current_earnings;
 	int current_score;
 public:
